refactor: moved all supabase utils and modals to one unified folder
Relates #166
