+++ app/(dashboard)/add-item/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 import ButtonRounded from '@/components/buttons/ButtonRounded';
-import insertRow from '@/utils/supabase/insertRow';
+import insertRow from '@/supabase/modals/insertRow';
 import { useForm } from 'react-hook-form';
 import { PartialItem } from '@/types/types';
 import UploadImageInput from '@/components/form/UploadImageInput';
