+++ middleware.ts
@@ -1,5 +1,5 @@
 import { NextResponse, type NextRequest } from 'next/server';
-import { createClient } from '@/utils/supabase/middleware';
+import { createClient } from '@/supabase/modals/middleware';
 
 export async function middleware(request: NextRequest) {
   try {
