+++ supabase/modals/getItems.ts
@@ -1,4 +1,4 @@
-import newClient from '../../config/createClient';
+import newClient from '../utils/createClient';
 import { PartialItem } from '../../types/types';
 
 export async function getItems(
